Move saved search functions to user_management.py
